+++ subprojects/platform-jvm/src/main/java/org/gradle/jvm/tasks/api/internal/ApiMemberSelector.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.jvm.tasks.api;
+package org.gradle.jvm.tasks.api.internal;
 
 import com.google.common.collect.Sets;
 
@@ -30,11 +30,11 @@ import static org.objectweb.asm.Opcodes.*;
 
 /**
  * Visits each {@link Member} of a given class and selects only those members that should
- * be included in the {@link ApiJar}. Selected API members are delegated to an adapter that
- * determines how to further process those members (e.g. stripping out method bodies), and
- * how to write a new "API class" with them.
+ * be included in the {@link org.gradle.jvm.tasks.api.ApiJar}. Selected API members are
+ * delegated to an adapter that determines how to further process those members (e.g.
+ * stripping out method bodies), and how to write a new "API class" with them.
  */
-class ApiMemberSelector extends ClassVisitor {
+public class ApiMemberSelector extends ClassVisitor {
 
     private final SortedSet<MethodMember> methods = Sets.newTreeSet();
     private final SortedSet<FieldMember> fields = Sets.newTreeSet();
