+++ subprojects/platform-jvm/src/main/java/org/gradle/jvm/tasks/api/internal/SimpleAnnotationValue.java
@@ -14,26 +14,11 @@
  * limitations under the License.
  */
 
-package org.gradle.jvm.tasks.api;
+package org.gradle.jvm.tasks.api.internal;
 
-import org.objectweb.asm.Type;
+public class SimpleAnnotationValue extends AnnotationValue<Object> {
 
-import java.lang.reflect.Modifier;
-
-class FieldMember extends TypedMember implements Comparable<FieldMember> {
-
-    public FieldMember(int access, String name, String signature, String typeDesc) {
-        super(access, name, signature, typeDesc);
-    }
-
-    @Override
-    public int compareTo(FieldMember o) {
-        return super.compare(o).result();
-    }
-
-    @Override
-    public String toString() {
-        return String.format(
-            "%s %s %s", Modifier.toString(getAccess()), Type.getType(getTypeDesc()).getClassName(), getName());
+    public SimpleAnnotationValue(String name, Object value) {
+        super(name, value);
     }
 }
